Switch to elisp--preceding-sexp to get the form to expand
This is sort of naughty because I'm using an internal function to get the
expression. On the other hand thing-at-point was failing me in some
situations and the use I have for expando-macro is one where it's just like
C-x C-e but it expands a form rather than executes it.

So..... although it's naughty to use an internal I'm going to do it anyway.